rename api.js to v1.js, since that's how it's imported
