+++ src/main.js
@@ -6,7 +6,7 @@ let monitor           = require('taskcluster-lib-monitor');
 let validator         = require('taskcluster-lib-validate');
 let docs              = require('taskcluster-lib-docs');
 let _                 = require('lodash');
-let v1                = require('./api');
+let v1                = require('./v1');
 let taskcluster       = require('taskcluster-client');
 let data              = require('./data');
 let RabbitAlerter     = require('./rabbitalerter');
